Task4 uint modified to uint8_t for base

+++ Lesson4/task4/src/main.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 //using namespace std;
 
-std::vector<int> plus(const std::vector<int> &A, const std::vector<int> &B, uint base)
+std::vector<int> plus(const std::vector<int> &A, const std::vector<int> &B, uint8_t base)
 {
     std::vector<int> C;
     std::vector<int>::const_reverse_iterator itA = A.rbegin();
@@ -32,7 +32,7 @@ std::vector<int> plus(const std::vector<int> &A, const std::vector<int> &B, uint
     return C;
 }
 
-std::pair<std::vector<int>,int> div(const std::vector<int> &A, uint baseIn, uint divider)
+std::pair<std::vector<int>,int> div(const std::vector<int> &A, uint8_t baseIn, uint8_t divider)
 {
     std::vector<int> C;
     std::vector<int>::const_iterator it = A.begin();
@@ -68,7 +68,7 @@ void printVector(const std::vector<int> A)
     std::cout << std::endl;
 }
 
-std::vector<int> convertBase(const std::vector<int> A, uint baseIn, uint baseOut)
+std::vector<int> convertBase(const std::vector<int> A, uint8_t baseIn, uint8_t baseOut)
 {
     std::vector<int> C;
     std::vector<int> temp;
@@ -85,7 +85,7 @@ std::vector<int> convertBase(const std::vector<int> A, uint baseIn, uint baseOut
     return C;
 }
 
-std::vector<int> convertBase2(const std::vector<int> A, uint baseIn, uint baseOut)
+std::vector<int> convertBase2(const std::vector<int> A, uint8_t baseIn, uint8_t baseOut)
 {
     std::vector<int> C;
     std::vector<int> t = A;
@@ -101,7 +101,7 @@ std::vector<int> convertBase2(const std::vector<int> A, uint baseIn, uint baseOu
     return C;
 }
 
-std::vector<int> plus(const std::vector<int> &A, uint baseA, const std::vector<int> &B, uint baseB, int baseResult)
+std::vector<int> plus(const std::vector<int> &A, uint8_t baseA, const std::vector<int> &B, uint8_t baseB, int baseResult)
 {
     return plus(convertBase(A, baseA, baseResult), convertBase(B, baseB, baseResult), baseResult);
 }
@@ -110,9 +110,9 @@ int main(int argc, char** args)
 {
     enum n {A=10,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P,R,S,T,U,V,W,X,Y,Z};
     std::vector<int> vA {F,F};
-    uint baseA = 16;
+    uint8_t baseA = 16;
     std::vector<int> vB {H,E,L,L,O};
-    uint baseB = 36;
+    uint8_t baseB = 36;
 
     std::cout << "Number A in a positional base "<< baseA <<" numeral system:" << std::endl;
     printVector(convertBase2(vA, baseA, baseA));
